requirements.txt
python-nmap
google
beatifulsoap4
requests